+++ scripts/render-markdown.js
@@ -3,60 +3,133 @@ import { gfm } from "./lib/turndown-plugin-gfm.js";
 
 let turndownService;
 
-export async function convertHtmlToMarkdown(html) {
-//https://deepwiki.com/mixmark-io/turndown/1-overview
-// https://github.com/trutohq/turndown-plugin-gfm
-//https://unpkg.com/@truto/turndown-plugin-gfm@1.0.2/lib/index.js
-
-//This was adapted and updated to newer version thanks to the code of Farling
-//https://github.com/farling42/fvtt-export-markdown/tree/master
-
-/*
-//We do not use doc, we only parse the html
-//export async function convertHtml(doc, html) {
-    // Foundry uses "showdown" rather than "turndown":
-    // SHOWDOWN fails to parse tables at all
-*/
-    if (!turndownService) {
-        // Setup Turndown service to use GFM for tables
-        // headingStyle: "atx"  <-- use "###" to indicate heading (whereas setext uses === or --- to underline headings)
+export async function convertHtmlToMarkdown(html, calloutSettings) {
+    
+    // We need to create the objekt each time new because the settings might have changed meanwhile and thus the rules also will change
+    //if (!turndownService) {
         turndownService = new TurndownService({
             headingStyle: "atx",
             codeBlockStyle: "fenced",
             bulletListMarker: "-"
         });
-        // GFM provides supports for strikethrough, tables, taskListItems, highlightedCodeBlock
-        //gfm = TurndownPluginGfmService.gfm;
         
-        //We imported gfm directly as we use a slightly different plugin
         turndownService.use(gfm);
+
+        // Rule for details tag callout type
+    if(calloutSettings.isRenderNoteCallout){
+        turndownService.addRule('obsidianDetailsCallout', {
+            filter: 'details',
+            replacement: ObsidianNoteCallout 
+        });
+        }
+
+        // Rule for Blockquote tag
+         if(calloutSettings.isRenderInfoCallout){
+        turndownService.addRule('obsidianInfoCallout', {
+             filter: 'blockquote',
+            replacement: ObsidianInfoCallout
+        });
     }
+
+        //Rule for section tag callout type which is a secret in Foundry     
+        if(calloutSettings.isRenderSecretCallout){
+        // Dirty fix so summary tag does not get dropped to soon ...
+        // TODO: Try do do it with keep rule in the next iteration
+        turndownService.addRule('ignoreSummary', {
+            filter: 'summary',
+            replacement: () => '' 
+        });
+        
+        turndownService.addRule('obsidianSecretCallout', {
+             filter: function (node, options) {
+                console.log("QJPC: Secret Node info ",node)
+                console.log("QJPC: Secret Node NAME ",node.nodeName)
+                console.log("QJPC: Secret Node CLASS ",node.classList)
+                return (node.nodeName === 'SECTION' && node.classList.contains('secret'))
+            },
+
+            replacement: ObsidianSecretCallout
+        });}
+    //} end of if statement for check of Turndown object
+
     let markdown;
     try {
-
-    	//We do not convert links as we only export single pages or journals
-    	//also secrets have been allready handled in the DOM manipulation and are allready removed in the html
-
-    	/*
-        // Convert links BEFORE doing HTML->MARKDOWN (to get links inside tables working properly)
-        // The conversion "escapes" the "[[...]]" markers, so we have to remove those markers afterwards
-       
-        //const part1 = convertLinks(html, doc)
-        //const include_gm = game.settings.get(MOD_CONFIG.MODULE_NAME, MOD_CONFIG.OPTION_INCLUDE_GM_ONLY);
-        //const part2 = await foundry.applications.ux.TextEditor.implementation.enrichHTML(part1, { secrets: include_gm });
-        //markdown = turndownService.turndown(part2).replaceAll("\\[\\[", "[[").replaceAll("\\]\\]", "]]");
-    	*/
-
         markdown = turndownService.turndown(html).replaceAll("\\[\\[", "[[").replaceAll("\\]\\]", "]]");
-
-        /*
-        // Now convert file references: inside a table it will be "\_" rather than just "/"
-        //const filepattern = /!\[\]\(([^)]*)\)/g;
-        //markdown = markdown.replaceAll(filepattern, replaceLinkedFile);
-    	*/
-
+        //markdown = markdown.replaceAll(/\n\n>\s*\[/g, '\n\n\n> [');
     } catch (error) {
-        console.warn(`Error: failed to decode html:`, html)
+        console.warn(`Error: failed to decode html:`, html, error);
     }
     return markdown;
+}
+
+// Function for Turndown rule
+function ObsidianNoteCallout(content, node) {
+    // Get the title from the summary tag
+    const summaryElement = node.querySelector('summary');
+    //const title = summaryElement ? summaryElement.textContent.trim() : 'Note';
+    const title = summaryElement ? turndownService.turndown(summaryElement).trim() : 'Note';
+    
+    // Create the DOM element and create a clone
+    // Reomove the summary tag so the remainder can be parsed to markdown
+    const parser = new DOMParser();
+    const doc = parser.parseFromString(node.innerHTML, 'text/html');
+    const tempSummary = doc.querySelector('summary');
+    if (tempSummary) {
+        tempSummary.remove();
+    }
+    
+    // Turndown is a global objekt so we can access it here
+    // Parse the remaining detials tag node to Markdown
+    const innerContent = turndownService.turndown(doc.body.innerHTML).trim();
+    
+    // Split the lines and add a ">" in front also empty line need to receive a ">"
+    const formattedContent = innerContent
+      .split('\n')
+      .map(line => line.trim() ? `> ${line}` : '>')
+      .join('\n')
+      .trimEnd();
+      
+    // Return the Callout
+    return `\n\n> [!note]- ${title}\n${formattedContent}\n\n`;
+}
+
+function ObsidianInfoCallout(content, node) {
+ 
+    const parser = new DOMParser();
+    const doc = parser.parseFromString(node.innerHTML, 'text/html');
+ 
+    const innerContent = turndownService.turndown(doc.body.innerHTML).trim();
+    
+    const formattedContent = innerContent
+      .split('\n')
+      .map(line => line.trim() ? `> ${line}` : '>')
+      .join('\n');
+      
+    const  finalMarkdown = `\n\n> [!info]- Info\n${formattedContent}\n\n`
+
+    return finalMarkdown;
+}
+
+function ObsidianSecretCallout(content,node){
+    
+    //Get the class and find out if it is a revealed or closed secret
+    let secretType =""
+    if(node.classList.contains('revealed')) {
+        secretType="Revealed"
+    } else {
+        secretType ="Hidden"
+    }
+
+    const parser = new DOMParser();
+    const doc = parser.parseFromString(node.innerHTML, 'text/html');
+  
+    const innerContent = turndownService.turndown(doc.body.innerHTML).trim();
+    const formattedContent = innerContent
+      .split('\n')
+      .map(line => line.trim() ? `> ${line}` : '>')
+      .join('\n');
+      
+
+    return `\n\n> [!secret]- ${secretType} Secret\n${formattedContent}\n\n`;
+
 }